I am making cool changes
This change was very important to the infrastructure. In order to accomodate the schema changes we had made, and in order for our REST api to corresond we made this change. We also made this change in order for our REST api to correspond accurately. Added a TODO that I will come back later and actually implement any of this.